added confirm windows for publish/delete actions

+++ views/admin/articleView.php
@@ -14,6 +14,7 @@
 		  });
 		</script>
 
+
     </head>
         
     <body>
@@ -42,7 +43,8 @@
 		        		<colgroup width="25%" />
 		        		<colgroup width="15%" />
 		        		<colgroup width="10%" />
-		        		<colgroup width="10%" />
+		        		<colgroup width="5%" />
+		        		<colgroup width="5%" />
 		        		<colgroup width="25%" />
 		        		<tr class="table-title">
 		        			<th>Date de publication</th>
@@ -61,14 +63,14 @@
 		        		<tr class="table-content">
 		        			<td><?php echo $articles[$i]->creation_date; ?></td>
 		        			<td><?php echo htmlspecialchars($articles[$i]->title); ?></td>
-		        			<td class="<?php echo $article[$i]->article_status; ?>-status"><span><?php echo $articles[$i]->article_status; ?></span></td>
+		        			<td class="<?php echo $articles[$i]->article_status; ?>-status"><span><?php echo $articles[$i]->article_status; ?></span></td>
 		        			<td><a class="btn-new-comment" href="index.php?action=adminGetArticle&id=<? echo $articles[$i]->id ?>"><?php echo $articles[$i]->nbNewComment; ?></a></td>
 		        			<td>Nb signalés</td>
 							<td><?php echo $articles[$i]->nbTotalComments; ?></td>
 		        			<td class="<?php echo $articles[$i]->article_status; ?>" >
-		        				<a class="btn-publish" href="index.php?action=publishArticle&id=<?php echo $articles[$i]->id ?>">Publier |</a>  
+		        				<a class="btn-publish" onClick="javascript: return confirm('Voulez-vous publier cet article ?');" href="index.php?action=publishArticle&id=<?php echo $articles[$i]->id ?>">Publier |</a>  
 								<a class="btn-modify" href="index.php?action=editArticle&id=<?php echo $articles[$i]->id ?>">Modifier</a> | 
-								<a class="btn-delete" href="index.php?action=deleteArticle&id=<?php echo $articles[$i]->id ?>">Supprimer</a>
+								<a class="btn-delete" onClick="javascript: return confirm('Voulez-vous vraiment supprimer cet article ? Cette action est définitive');" href="index.php?action=deleteArticle&id=<?php echo $articles[$i]->id ?>">Supprimer</a>
 							</td>
